Update jobs and zos_job_submit to better support jobs in the input queue
